Implemented parent git directory detection
The git detection just looked if in the current working directory a .git
folder was located. If so, an internal flag was set but the information
was never used.

Now the shell context also checks if one of the parent directories is a
git diretory, stopping when hitting the user home. If so, for now (git)
is rendered at the end of the prompt to indicate a git repository.

In one of the next steps the branch and some meta information should be
added to the prompt and also made configurable.