Quick fix for running lsh in wsl2
For some reason, the path indexing and getting the username in wsl
is not working due to filesystem issues. For now I introduced
a build flag `EnableWslCompatibility` that disables the functionality
that is causing the issue with wsl. This is a temporary fix, and should
be investigated and fixed properly.