Fixed indexing issue related to wsl
Running in wsl, lake-sh tried to also index windows mounts, which
led to issues. Now indexing is restricted to the unix filesystem (not
accessing /mnt), which resolves the issue.

Being unable to index at all, disabled all external commands in lsh,
so with resolving this issue, lsh is now usable again in context of wsl.