Fixed alias loading from profile
It is possible to configure an alias for a command in the lake shell
profile (.lakesh) located in the home directory. The resolved alias
where never used, so using an alias like ll was never translated to ls.
Now there is a check in place, that takes an alias into consideration
if it could be resolved.