Implemented concat support via &&
Now it's possible to concat multiple commands using && like in zsh or
bash. The implementation is a quick & dirty parser hack, that just
checks if the input string contains && and sets a flag indicating down
the line, that commands are concat and not piped.

This is by no means a solid long term solution, but will make the shell
more usable on a daily basis.