:art: Adding `WorkingDirectory` type
Adding dedicated type for representing working directories, to have a
better separation of concerns. The new `WorkingDirectory` type manages
the path of the current working directoy and is also responsible to
check the attributes of the current directoy, which currently only
involves checking if the current dir is a git repository or not.